requirements.txt
typing==3.7.4.3
treys==0.1.8
markdown
google-adk
langchain-core

# LLM Dependencies for poker transcript processing
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
litellm>=1.57.0
langchain-community>=0.3.0

# Optional: for local LLM support
# ollama>=0.1.0
# transformers>=4.30.0
# torch>=2.0.0